requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
coverage==7.6.9
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
